requirements.txt
requests
pyyaml
pandas
networkx
matplotlib
jinja2